+++ rspack.config.ts
@@ -52,7 +52,7 @@ export default {
     service_worker: `${src}/service_worker.ts`,
     offscreen: `${src}/offscreen.ts`,
     sandbox: `${src}/sandbox.ts`,
-    ff_persistent: `${src}/ff_persistent.ts`,
+    persistent_frame: `${src}/persistent_frame.ts`,
     content: `${src}/content.ts`,
     scripting: `${src}/scripting.ts`,
     inject: `${src}/inject.ts`,
@@ -235,11 +235,11 @@ export default {
       chunks: ["sandbox"],
     }),
     new rspack.HtmlRspackPlugin({
-      filename: `${dist}/ext/src/ff_persistent.html`,
-      template: `${src}/pages/ff_persistent.html`,
+      filename: `${dist}/ext/src/persistent_frame.html`,
+      template: `${src}/pages/persistent_frame.html`,
       inject: "head",
       minify: true,
-      chunks: ["ff_persistent"],
+      chunks: ["persistent_frame"],
     }),
     new ZipExecutionPlugin(),
   ].filter(Boolean),
