+++ src/service_worker.ts
@@ -8,7 +8,7 @@ import { MessageQueue } from "@Packages/message/message_queue";
 import { ServiceWorkerMessageSend } from "@Packages/message/window_message";
 import migrate, { migrateChromeStorage } from "./app/migrate";
 import { cleanInvalidKeys } from "./app/repo/resource";
-import { keepEventPageRunning } from "./persistent";
+import { keepEventPageRunning } from "@App/pkg/utils/persistent";
 
 migrate();
 migrateChromeStorage();
